Depth is now a short.

+++ src/main/java/it/albertus/eqbulletin/service/geofon/html/transformer/HtmlTableDataTransformer.java
@@ -101,7 +101,7 @@ public class HtmlTableDataTransformer {
 				longitude *= -1;
 			}
 
-			final int depth = Integer.parseInt(lines[4].substring(lines[4].indexOf(magnitudePrefix) + magnitudePrefix.length(), lines[4].indexOf(depthSuffix)).trim());
+			final short depth = Short.parseShort(lines[4].substring(lines[4].indexOf(magnitudePrefix) + magnitudePrefix.length(), lines[4].indexOf(depthSuffix)).trim());
 			final Status status = Status.valueOf(lines[5].substring(lines[5].lastIndexOf(statusPrefix) + statusPrefix.length(), lines[5].indexOf(statusSuffix)).trim());
 			final String region = lines[6].substring(lines[6].lastIndexOf(regionPrefix) + regionPrefix.length(), lines[6].lastIndexOf(regionSuffix)).trim();
